snd_PlayerHeap OK

+++ src/nw4r/snd/snd_MmlParser.cpp
@@ -159,7 +159,7 @@ MmlSeqTrack::ParseResult MmlParser::Parse(MmlSeqTrack *track,
 		s32 commandArg1 = 0;
 		s32 commandArg2 = 0;
 
-		switch (static_cast<int>(cmd & MML_CMD_SET_MASK))
+		switch (cmd & MML_CMD_SET_MASK)
 		{
 		case 0x80:
 		{
